Move confirmedReservationsForCapacity to ReservationRepository
The query's root model is Reservation, so it belongs in
ReservationRepository, not TableRepository. This fixes a cross-domain
leak where the Table repository owned a Reservation query, and is
symmetric with findAvailable() (root Table) staying put.
