+++ app/Repositories/TableRepository.php
@@ -51,14 +51,6 @@ class TableRepository
         return Table::matchingCapacity($seatsRequested)->count();
     }
 
-    public function confirmedReservationsForCapacity(int $seatsRequested, string $date): \Illuminate\Support\Collection
-    {
-        return Reservation::confirmed()
-            ->where('date', $date)
-            ->whereIn('table_id', Table::matchingCapacity($seatsRequested)->select('id'))
-            ->get(['table_id', 'start_time', 'end_time']);
-    }
-
     public function findAvailable(int $seatsRequested, string $date, string $startTime, string $endTime): Collection
     {
         return Table::matchingCapacity($seatsRequested)
